requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
openai==1.40.6
pandas==2.2.2
gspread==6.1.4
google-auth==2.31.0
python-dateutil==2.9.0.post0
pytz==2024.1
